Safari-609.1.11.1/LayoutTests/http/tests/referrer-policy-script/no-referrer/cross-origin-http-http-expected.txt

Tests the behavior of no-referrer referrer policy when cross origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
